(NPUP-15) Port right shift operator to the dispatcher.
Porting the right shift operator's implementation to the dispatcher.

This also removes the previous implementation of invoking binary operators.
