+++ lib/include/puppet/compiler/evaluation/operators/binary/right_shift.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file
+ * Declares the right shift operator.
+ */
+#pragma once
+
+#include "descriptor.hpp"
+
+namespace puppet { namespace compiler { namespace evaluation { namespace operators { namespace binary {
+
+    /**
+     * Implements the right shift operator.
+     */
+    struct right_shift
+    {
+        /**
+         * Create a binary operator descriptor.
+         * @return Returns the binary operator descriptor representing this operator.
+         */
+        static descriptor create_descriptor();
+    };
+
+}}}}}  // namespace puppet::compiler::evaluation::operators::binary
