--- lib/include/puppet/compiler/evaluation/operators/right_shift.hpp
@@ -1,24 +1,0 @@
-/**
- * @file
- * Declares the right shift operator.
- */
-#pragma once
-
-#include "binary_operator_context.hpp"
-
-namespace puppet { namespace compiler { namespace evaluation { namespace operators {
-
-    /**
-     * Implements the right shift operator.
-     */
-    struct right_shift
-    {
-        /**
-         * Called to invoke the operator.
-         * @param context The binary operator context.
-         * @return Returns the resulting value.
-         */
-        runtime::values::value operator()(binary_operator_context const& context) const;
-    };
-
-}}}}  // namespace puppet::compiler::evaluation::operators
